Updated Ubuntu 8.04 instructions Modified the update scripts to check if a download of the tar.gz has succeeded.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/trunk@1226 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/scripts/update_from_tgz.sh
@@ -8,8 +8,15 @@ rm -f ISPConfig-3-stable.tar.gz
 fi
 
 wget http://www.ispconfig.org/downloads/ISPConfig-3-stable.tar.gz
-tar xvfz ISPConfig-3-stable.tar.gz
-cd ispconfig3_install/install/
-php -q update.php
-rm -rf /tmp/ispconfig3_install/install
-rm -f ISPConfig-3-stable.tar.gz
+if [ -f ISPConfig-3-stable.tar.gz ]
+then
+	tar xvfz ISPConfig-3-stable.tar.gz
+	cd ispconfig3_install/install/
+	php -q update.php
+	rm -rf /tmp/ispconfig3_install/install
+	rm -f ISPConfig-3-stable.tar.gz
+else
+	echo "Unable to download the update."
+fi
+
+exit 0
